requirements.txt
discord.py>=2.3.0
instaloader>=4.10
